requirements.txt
torch>=2.7.1
torchvision>=0.14.0
timm>=0.6.13
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.2
scikit-learn>=1.0.0
Pillow>=9.0.0
opencv-python>=4.5.5
